+++ src/helpers/url/url.test.ts
@@ -6,14 +6,14 @@ interface RequestWithDocument extends Request {
   document: {
     path: string;
   };
+  apiPath: string;
 }
 
 describe('Url', () => {
   it('should get the url of a document', () =>
     expect(
       getUrl({
-        protocol: 'http',
-        headers: { host: 'localhost:8080' },
+        apiPath: 'http://localhost:8080',
         document: { path: '/news' },
       } as RequestWithDocument),
     ).toBe('http://localhost:8080/news'));
@@ -21,8 +21,7 @@ describe('Url', () => {
   it('should get the url of the root', () =>
     expect(
       getUrl({
-        protocol: 'http',
-        headers: { host: 'localhost:8080' },
+        apiPath: 'http://localhost:8080',
         document: { path: '/' },
       } as RequestWithDocument),
     ).toBe('http://localhost:8080'));
@@ -30,8 +29,7 @@ describe('Url', () => {
   it('should get the root url', () =>
     expect(
       getRootUrl({
-        protocol: 'http',
-        headers: { host: 'localhost:8080' },
-      } as Request),
+        apiPath: 'http://localhost:8080',
+      } as RequestWithDocument),
     ).toBe('http://localhost:8080'));
 });